new cli design start

+++ src/nodefony/core/cli/angular/angularCli.js
@@ -108,9 +108,10 @@ let angularCli = class angularCli extends nodefony.Service {
   cleanTmp() {
     try {
       let tmpDir = path.resolve(this.tmp, this.bundleName);
-      this.cli.existsSync(tmpDir);
       try {
-        shell.rm('-rf', tmpDir);
+        if (this.cli.existsSync(tmpDir)) {
+          shell.rm('-rf', tmpDir);
+        }
       } catch (e) {
         this.logger(e, "ERROR");
         throw e;
